# server/src/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bloom Smart Habit REST API",
    "version": "1.3.0",
    "description": "Versioned API contract for web and native mobile clients."
  },
  "servers": [
    {
      "url": "/v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/habits": {
      "get": {
        "operationId": "listHabits",
        "responses": {
          "200": {
            "description": "Current user's habits.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Habit"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createHabit",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHabitInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created habit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Habit"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/today": {
      "get": {
        "operationId": "getToday",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Habits due for the selected local date.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/Habit"
                      },
                      {
                        "type": "object",
                        "properties": {
                          "todayLog": {
                            "anyOf": [
                              {
                                "$ref": "#/components/schemas/HabitLog"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          }
                        }
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
    "/habits/{habitId}/logs/{localDate}": {
      "put": {
        "operationId": "checkInHabit",
        "parameters": [
          {
            "$ref": "#/components/parameters/HabitId"
          },
          {
            "name": "localDate",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckInInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Idempotent daily log.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HabitLog"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "delete": {
        "operationId": "removeHabitCheckIn",
        "parameters": [
          {
            "$ref": "#/components/parameters/HabitId"
          },
          {
            "name": "localDate",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Daily check-in removed.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "deleted"
                  ],
                  "properties": {
                    "deleted": {
                      "type": "boolean",
                      "const": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/journal/{localDate}": {
      "get": {
        "operationId": "getDailyJournal",
        "parameters": [
          {
            "name": "localDate",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Daily reflection for the selected date.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyJournal"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "put": {
        "operationId": "saveDailyJournal",
        "parameters": [
          {
            "name": "localDate",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DailyJournalInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Daily reflection saved.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyJournal"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/tracking": {
      "get": {
        "operationId": "getTrackingReport",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tracking report for a range of up to 366 days.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackingReport"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "OpaqueSessionToken"
      }
    },
    "parameters": {
      "HabitId": {
        "name": "habitId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 200
        }
      }
    },
    "responses": {
      "Error": {
        "description": "Consistent error envelope.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    },
    "schemas": {
...
      "Frequency": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "kind"
            ],
            "properties": {
              "kind": {
                "const": "daily"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "kind",
              "target"
            ],
            "properties": {
              "kind": {
                "const": "weekly_target"
              },
              "target": {
                "type": "integer",
                "minimum": 1,
                "maximum": 7
              }
            }
          },
          {
            "type": "object",
            "required": [
              "kind",
              "days"
            ],
            "properties": {
              "kind": {
                "const": "weekdays"
              },
              "days": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 6
                },
                "minItems": 1
              }
            }
          }
        ]
      },
...
      "CreateHabitInput": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/TemplateHabitInput"
          },
          {
            "$ref": "#/components/schemas/CustomHabitInput"
          }
        ]
      },
      "TemplateHabitInput": {
        "type": "object",
        "required": [
          "templateId"
        ],
        "properties": {
          "templateId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "CustomHabitInput": {
        "type": "object",
        "required": [
          "name",
          "icon",
          "category",
          "type",
          "frequency"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "icon": {
            "type": "string",
            "minLength": 1,
            "maxLength": 8
          },
          "category": {
            "type": "string",
            "enum": [
              "diet",
              "prayer",
              "steps",
              "gym",
              "food",
              "learning",
              "other"
            ]
          },
          "type": {
            "type": "string",
            "enum": [
              "do",
              "avoid",
              "count",
              "duration"
            ]
          },
          "target": {
            "type": [
              "number",
              "null"
            ],
            "exclusiveMinimum": 0
          },
          "unit": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 24
          },
          "frequency": {
            "$ref": "#/components/schemas/Frequency"
          },
          "forgiving": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "CheckInInput": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "done",
              "skipped",
              "partial"
            ]
          },
          "value": {
            "type": [
              "number",
              "null"
            ],
            "minimum": 0
          },
          "note": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 1000
          },
          "prayerStatus": {
            "type": [
              "string",
              "null"
            ],
            "enum": [
              "on_time",
              "late",
              "missed",
              null
            ]
          }
        }
      },
      "Habit": {
        "type": "object",
        "required": [
          "id",
          "user_id",
          "name",
          "icon",
          "category",
          "habit_type",
          "frequency",
          "state"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "habit_type": {
            "type": "string",
            "enum": [
              "do",
              "avoid",
              "count",
              "duration"
            ]
          },
          "target": {
            "type": [
              "number",
              "null"
            ]
          },
          "unit": {
            "type": [
              "string",
              "null"
            ]
          },
          "frequency": {
            "$ref": "#/components/schemas/Frequency"
          },
          "state": {
            "type": "string",
            "enum": [
              "active",
              "paused",
              "archived"
            ]
          },
          "reminder_enabled": {
            "type": "boolean"
          },
          "reminder_time": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "additionalProperties": true
      },
...
      "HabitLog": {
        "type": "object",
        "required": [
          "id",
          "habit_id",
          "user_id",
          "local_date",
          "status"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "habit_id": {
            "type": "string",
            "format": "uuid"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          },
          "local_date": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string",
            "enum": [
              "done",
              "skipped",
              "partial"
            ]
          },
          "value": {
            "type": [
              "number",
              "null"
            ]
          },
          "note": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "additionalProperties": true
      },
      "DailyJournalInput": {
        "type": "object",
        "properties": {
          "winNote": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 1000
          },
          "reflectionNote": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 1000
          }
        }
      },
      "DailyJournal": {
        "type": "object",
        "required": [
          "id",
          "user_id",
          "local_date",
          "win_note",
          "reflection_note"
        ],
        "properties": {
          "id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          },
          "local_date": {
            "type": "string",
            "format": "date"
          },
          "win_note": {
            "type": [
              "string",
              "null"
            ]
          },
          "reflection_note": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "TrackingHabit": {
        "type": "object",
        "required": [
          "id",
          "name",
          "icon",
          "habit_type",
          "target",
          "unit",
          "status",
          "value",
          "note"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "habit_type": {
            "type": "string",
            "enum": [
              "do",
              "avoid",
              "count",
              "duration"
            ]
          },
          "target": {
            "type": [
              "number",
              "null"
            ]
          },
          "unit": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "done",
              "skipped",
              "partial",
              "not_checked"
            ]
          },
          "value": {
            "type": [
              "number",
              "null"
            ]
          },
          "note": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "TrackingDay": {
        "type": "object",
        "required": [
          "date",
          "completed",
          "scheduled",
          "completionRate",
          "winNote",
          "reflectionNote",
          "habits"
        ],
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "completed": {
            "type": "integer",
            "minimum": 0
          },
          "scheduled": {
            "type": "integer",
            "minimum": 0
          },
          "completionRate": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100
          },
          "winNote": {
            "type": [
              "string",
              "null"
            ]
          },
          "reflectionNote": {
            "type": [
              "string",
              "null"
            ]
          },
          "habits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TrackingHabit"
            }
          }
        }
      },
      "TrackingReport": {
        "type": "object",
        "required": [
          "from",
          "to",
          "totalCompleted",
          "totalScheduled",
          "completionRate",
          "days"
        ],
        "properties": {
          "from": {
            "type": "string",
            "format": "date"
          },
          "to": {
            "type": "string",
            "format": "date"
          },
          "totalCompleted": {
            "type": "integer",
            "minimum": 0
          },
          "totalScheduled": {
            "type": "integer",
            "minimum": 0
          },
          "completionRate": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100
          },
          "days": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TrackingDay"
            }
          }
        }
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "code",
          "message",
          "correlationId",
          "retryable"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "correlationId": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          },
          "fieldErrors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
...
    }
  }
}